+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Updated/Enhanced Rules and Bug Fixes
  - #571: Support Groovy 3.0.6 and fix test failures.
  - #576: Ignore pointcuts for empty methods. ([Thomas Lefèvre](https://github.com/tlefevre))
  - #562: Deprecate the **UnnecessarySubstring** rule. Its recommendation to use subscripts on strings is not always safe/valid.
+ - #578: Add ability for **UnusedPrivateFieldRule** to raise violations for fields which are only used within constructors. ([Marcin Erdmann](https://github.com/erdi))
 
 Build and Infrastructure
  - #570: Publish a fat jar. ([Eric Citaire](https://github.com/ericcitaire))